+++ app/assets/javascripts/bali/controllers/dynamic-fields-controller.js
@@ -7,7 +7,7 @@ import {
   nextSibling
 } from '../utils/domHelpers.js'
 
-// TODO: Add tests (Issue: #155)
+// Contract frozen by cypress/e2e/dynamic-fields-controller.cy.js (#155/#715).
 
 export class DynamicFieldsController extends Controller {
   static targets = ['template', 'container', 'button']